+++ src/components/v5/common/ActionSidebar/partials/forms/CreateNewTeamForm/hooks.ts
@@ -4,13 +4,13 @@ import { useWatch } from 'react-hook-form';
 import { type DeepPartial } from 'utility-types';
 
 import { useColonyContext } from '~context/ColonyContext/ColonyContext.ts';
-import { ActionTypes } from '~redux/index.ts';
-import { DecisionMethod } from '~types/actions.ts';
+import { type DecisionMethod } from '~types/actions.ts';
 import { mapPayload, pipe } from '~utils/actions.ts';
 import { DECISION_METHOD_FIELD_NAME } from '~v5/common/ActionSidebar/consts.ts';
 
 import useActionFormBaseHook from '../../../hooks/useActionFormBaseHook.ts';
 import { type ActionFormBaseProps } from '../../../types.ts';
+import { getDomainFormActionType } from '../utils/getDomainFormActionType.ts';
 
 import { validationSchema, type CreateNewTeamFormValues } from './consts.ts';
 import { getCreateNewTeamPayload } from './utils.tsx';
@@ -24,10 +24,7 @@ export const useCreateNewTeam = (
   });
 
   useActionFormBaseHook({
-    actionType:
-      decisionMethod === DecisionMethod.Permissions
-        ? ActionTypes.ACTION_DOMAIN_CREATE
-        : ActionTypes.MOTION_DOMAIN_CREATE_EDIT,
+    actionType: getDomainFormActionType(decisionMethod),
     defaultValues: useMemo<DeepPartial<CreateNewTeamFormValues>>(
       () => ({
         createdIn: Id.RootDomain,
