+++ src/components/v5/common/ActionSidebar/partials/forms/CreateNewTeamForm/utils.tsx
@@ -30,7 +30,6 @@ export const getCreateNewTeamPayload = (
     ...baseDomainPayload,
     colonyRoles: extractColonyRoles(colony.roles),
     colonyDomains: extractColonyDomains(colony.domains),
-    isMultiSig: values.decisionMethod === DecisionMethod.MultiSig,
     domainCreatedInNativeId: values.createdIn,
   };
 };
