+++ src/components/v5/common/ActionSidebar/partials/forms/EditTeamForm/hooks.ts
@@ -4,14 +4,14 @@ import { useFormContext, useWatch } from 'react-hook-form';
 import { type DeepPartial } from 'utility-types';
 
 import { useColonyContext } from '~context/ColonyContext/ColonyContext.ts';
-import { ActionTypes } from '~redux/index.ts';
-import { DecisionMethod } from '~types/actions.ts';
+import { type DecisionMethod } from '~types/actions.ts';
 import { mapPayload, pipe } from '~utils/actions.ts';
 import { findDomainByNativeId } from '~utils/domains.ts';
 
 import { DECISION_METHOD_FIELD_NAME } from '../../../consts.ts';
 import useActionFormBaseHook from '../../../hooks/useActionFormBaseHook.ts';
 import { type ActionFormBaseProps } from '../../../types.ts';
+import { getDomainFormActionType } from '../utils/getDomainFormActionType.ts';
 
 import { validationSchema, type EditTeamFormValues } from './consts.ts';
 import { getEditDomainPayload } from './utils.tsx';
@@ -45,10 +45,7 @@ export const useEditTeam = (
   useActionFormBaseHook({
     getFormOptions,
     validationSchema,
-    actionType:
-      decisionMethod === DecisionMethod.Permissions
-        ? ActionTypes.ACTION_DOMAIN_EDIT
-        : ActionTypes.MOTION_DOMAIN_CREATE_EDIT,
+    actionType: getDomainFormActionType(decisionMethod),
     defaultValues: useMemo<DeepPartial<EditTeamFormValues>>(
       () => ({
         createdIn: team || Id.RootDomain,
