+++ src/redux/actionTypes.ts
@@ -112,6 +112,12 @@ export enum ActionTypes {
   MOTION_DOMAIN_CREATE_EDIT = 'MOTION_DOMAIN_CREATE_EDIT',
   MOTION_DOMAIN_CREATE_EDIT_ERROR = 'MOTION_DOMAIN_CREATE_EDIT_ERROR',
   MOTION_DOMAIN_CREATE_EDIT_SUCCESS = 'MOTION_DOMAIN_CREATE_EDIT_SUCCESS',
+  MOTION_REPUTATION_DOMAIN_CREATE_EDIT = 'MOTION_REPUTATION_DOMAIN_CREATE_EDIT',
+  MOTION_REPUTATION_DOMAIN_CREATE_EDIT_ERROR = 'MOTION_REPUTATION_DOMAIN_CREATE_EDIT_ERROR',
+  MOTION_REPUTATION_DOMAIN_CREATE_EDIT_SUCCESS = 'MOTION_REPUTATION_DOMAIN_CREATE_EDIT_SUCCESS',
+  MOTION_MULTISIG_DOMAIN_CREATE_EDIT = 'MOTION_MULTISIG_DOMAIN_CREATE_EDIT',
+  MOTION_MULTISIG_DOMAIN_CREATE_EDIT_ERROR = 'MOTION_MULTISIG_DOMAIN_CREATE_EDIT_ERROR',
+  MOTION_MULTISIG_DOMAIN_CREATE_EDIT_SUCCESS = 'MOTION_MULTISIG_DOMAIN_CREATE_EDIT_SUCCESS',
   MOTION_EDIT_COLONY = 'MOTION_EDIT_COLONY',
   MOTION_EDIT_COLONY_ERROR = 'MOTION_EDIT_COLONY_ERROR',
   MOTION_EDIT_COLONY_SUCCESS = 'MOTION_EDIT_COLONY_SUCCESS',
